Revert "Auto-format (#473)"
This reverts commit ed85c0dc16db67c49ed192bdd335ca11849cddb8.

+++ src/main/java/com/paneedah/weaponlib/vehicle/smoothlib/QPTI.java
@@ -5,14 +5,14 @@ import com.paneedah.weaponlib.vehicle.jimphysics.InterpolationKit;
 import static com.paneedah.mwc.proxies.ClientProxy.MC;
 
 public class QPTI {
-
-    public static float pti(float a, float b) {
-        return (float) InterpolationKit.interpolateValue(a, b, MC.getRenderPartialTicks());
-    }
-
-    public static double pti(double a, double b) {
-        return InterpolationKit.interpolateValue(a, b, MC.getRenderPartialTicks());
-
-    }
+	
+	public static float pti(float a, float b) {
+		return (float) InterpolationKit.interpolateValue((double) a, (double) b, (double) MC.getRenderPartialTicks());
+	}
+	
+	public static double pti(double a, double b) {
+		return InterpolationKit.interpolateValue((double) a, (double) b, (double) MC.getRenderPartialTicks());
+		
+	}
 
 }
